requirements.txt
fastapi==0.118.3
uvicorn[standard]==0.34.0

python-multipart==0.0.20
sqlalchemy==2.0.36
databases==0.9.0

psycopg2-binary==2.9.11
pymysql==1.1.1

bcrypt==4.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

Pillow
numpy                    # 버전 고정 해제 (3.14 공식 지원 버전 나오면 그걸 설치)

psutil==7.0.0

pydantic==2.12.0
pydantic-settings==2.12.0

alembic==1.13.1
